refactor: move the state processing into a separate component ss-057
