+++ src/pages/game-selection-page/game-selection-page.tsx
@@ -1,9 +1,5 @@
-import { FallbackMessage, Loader } from "~/libs/components/components";
-import { EMPTY_ARRAY_LENGTH } from "~/libs/constants/constants";
-import { DataStatus } from "~/libs/enums/enums";
 import {
 	useAppDispatch,
-	useAppSelector,
 	useCallback,
 	useEffect,
 	useLanguageSync,
@@ -11,14 +7,12 @@ import {
 } from "~/libs/hooks/hooks";
 import { actions as gamesActions } from "~/modules/games/games";
 
-import { GameCard } from "./game-card";
+import { GameSelectionList } from "./game-selection-list";
 import styles from "./styles.module.css";
 
 const GameSelectionPage: React.FC = () => {
 	const { t } = useTranslation();
 	const dispatch = useAppDispatch();
-	const games = useAppSelector((state) => state.games.games);
-	const gamesStatus = useAppSelector((state) => state.games.gamesStatus);
 
 	const fetchGames = useCallback(() => {
 		void dispatch(gamesActions.getAllGames());
@@ -30,24 +24,6 @@ const GameSelectionPage: React.FC = () => {
 		fetchGames();
 	}, [fetchGames]);
 
-	const isLoading = gamesStatus === DataStatus.IDLE || gamesStatus === DataStatus.PENDING;
-	const isFailed = gamesStatus === DataStatus.REJECTED;
-	const isEmpty = gamesStatus === DataStatus.FULFILLED && games.length === EMPTY_ARRAY_LENGTH;
-
-	let content: React.ReactNode = null;
-
-	if (isLoading) {
-		content = <Loader hasOverlay />;
-	} else if (isFailed) {
-		content = <FallbackMessage message={t("games.selection.error")} />;
-	} else if (isEmpty) {
-		content = (
-			<div className={styles["game-selection-page__no-games"]}>{t("games.selection.empty")}</div>
-		);
-	} else {
-		content = games.map((game) => <GameCard game={game} key={game.id} />);
-	}
-
 	return (
 		<div className={styles["game-selection-page__container"]}>
 			<header className={styles["game-selection-page__header"]}>
@@ -55,7 +31,7 @@ const GameSelectionPage: React.FC = () => {
 			</header>
 
 			<main aria-live="polite" className={styles["game-selection-page__grid"]}>
-				{content}
+				<GameSelectionList />
 			</main>
 		</div>
 	);
